Separated curation engine out so it can be reuseed in other processes
